+++ server/models/admin.js
@@ -1,10 +1,10 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   const Admin = sequelize.define('Admin', {
-    ["a-id"]: DataTypes.INTEGER,
+    ["a-id"]: DataTypes.BIGINT,
     ["a-name"]: DataTypes.STRING,
     ["a-pwd"]: DataTypes.STRING,
-    ["a-token"]: DataTypes.INTEGER,
+    ["a-token"]: DataTypes.BIGINT,
     ["a-type"]: DataTypes.INTEGER,
   }, {});
   Admin.associate = function(models) {
