+++ server/models/teacher.js
@@ -1,10 +1,10 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   const Teacher = sequelize.define('Teacher', {
-    ["t-id"]: DataTypes.INTEGER,
+    ["t-id"]: DataTypes.BIGINT,
     ["t-name"]: DataTypes.STRING,
     ["t-pwd"]: DataTypes.STRING,
-    ["t-token"]: DataTypes.INTEGER,
+    ["t-token"]: DataTypes.BIGINT,
     ["t-type"]: DataTypes.INTEGER
   }, {});
   Teacher.associate = function(models) {
